requirements.txt
# OCR処理
pytesseract==0.3.10
Pillow==10.1.0
pdf2image==1.16.3

# Google Sheets API
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0

# PDF処理
PyPDF2==3.0.1

# データ処理
pandas==2.1.4
openpyxl==3.1.2

# ユーティリティ
python-dotenv==1.0.0


